% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffplot.R
\name{ffplot}
\alias{ffplot}
\alias{ffplot.funres}
\alias{ffplot.default}
\title{Function-function plot}
\usage{
ffplot(
  object,
  resolution = 101,
  n = NULL,
  ref.col = 2,
  ref.lwd = 1,
  ref.lty = "dashed",
  ...
)

\method{ffplot}{funres}(
  object,
  resolution = 101,
  n = NULL,
  ref.col = 2,
  ref.lwd = 1,
  ref.lty = "dashed",
  ...
)

\method{ffplot}{default}(
  object,
  resolution = 101,
  n = NULL,
  ref.col = 1,
  ref.lwd = 1,
  ref.lty = "dashed",
  ...
)
}
\arguments{
\item{object}{Either a \link[=fresiduals]{funres} object (i.e., a list of
functional residuals) or a fitted model for which the extraction of
\link[=fresiduals]{functional residuals} is meaningful (e.g., a \link[stats:glm]{glm}
object).}

\item{resolution}{Integer specifying the number of points between 0 and 1 to
use. Default is 101.}

\item{n}{Integer specifying the number of functional residuals to draw at
random, which can help with compute time. Default is \code{NULL} which corresponds
to no subsampling.}

\item{ref.col}{Character string or integer specifying the color to use for
the 45-degree reference line. Default is 1.}

\item{ref.lwd}{Integer specifying the line width to use for the 45-degree
reference line. Default is 1.}

\item{ref.lty}{Character string or integer specifying the line type to use
for the 45-degree reference line. Default is \code{"dashed"}.}

\item{...}{Additional optional arguments pass to \link[base:plot]{plot()}.}
}
\description{
Constructs a function-function (Fn-Fn) plot. That is, a plot of the average
functional residual against the CDF of a U(0, 1) random variable.
}
