% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredplot.R
\name{fredplot}
\alias{fredplot}
\title{Functional residual density plot}
\usage{
fredplot(
  object,
  x,
  resolution = 101,
  scale = c("uniform", "normal"),
  geom = c("hex", "kde", "bin"),
  ...
)
}
\arguments{
\item{object}{An object of class \link{fresiduals}.}

\item{x}{Vector of predictor values.}

\item{resolution}{Integer specifying the resolution of the plot. Default
is 101.}

\item{scale}{Character string specifying which scale to use for plotting.
Default is \code{"uniform"}.}

\item{geom}{Character string providing the type of \link[ggplot2:ggplot2-ggproto]{geom} to
plot. Current options include \code{"hex"} (the default) for a hexagonal heatmap
of 2D bin counts, \code{"kde"} for 2D kernel density estimation, or \code{"bin"} for a
heatmap of 2D bin counts.}

\item{...}{Additional optional arguments to be passed to the respective
\link[ggplot2:ggplot2-ggproto]{geom}.}
}
\description{
Generates a functional residual-vs-predictor plot, visualizing the density of
the functional residuals as a heatmap rendered on either the uniform or
normal scale.
}
\examples{
# Generate data from a logistic regression model with quadratic form
set.seed(1217)
n <- 10000
x <- rnorm(n)
z <- 1 - 2*x + 3*x^2 + rlogis(n)
y <- ifelse(z > 0, 1, 0)

# Fit models with/without quadratic term
fit.wrong <- glm(y ~ x, family = binomial)  # wrong
fit.right <- glm(y ~ x + I(x^2), family = binomial)  # right

library(ggplot2)
library(patchwork)
theme_set(theme_bw())

# Functional REsidual Density plot for each model
fredplot(fit.wrong, x = x) +
  scale_fill_distiller(palette = "Spectral") +
  ggtitle("Wrong model") +
  fredplot(fit.right, x = x) +
  scale_fill_distiller(palette = "Spectral") +
  ggtitle("Correct model")

}
