% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fresiduals.R
\name{fresiduals}
\alias{fresiduals}
\alias{fresiduals.default}
\alias{fresiduals.matrix}
\title{Function residuals}
\usage{
fresiduals(
  object,
  type = c("function", "surrogate", "probscale"),
  link.scale = TRUE,
  ...
)

\method{fresiduals}{default}(
  object,
  type = c("function", "surrogate", "probscale"),
  link.scale = TRUE,
  ...
)

\method{fresiduals}{matrix}(
  object,
  type = c("function", "surrogate", "probscale"),
  link.scale = TRUE,
  ...
)
}
\arguments{
\item{object}{A fitted model object. Currently supported model types include:
\itemize{
\item \link[stats:glm]{glm} - generalized linear models from the core \strong{stats}
package.
\item \link[mgcv:gam]{gam} - generalized additive models from the
\href{https://cran.r-project.org/package=mgcv}{mgcv} package.
\item \link[pscl:zeroinfl]{zeroinfl} - zero-inflated Poisson (ZIP) models from the
\href{https://cran.r-project.org/package=pscl}{pscl} package.
\item \link[VGAM:vglm]{vglm}/\link[VGAM:vgam]{vgam} - vector generalized linear/additive
models (VGLMs/VGAMs) from the \href{https://cran.r-project.org/package=VGAM}{VGAM}
package.
}}

\item{type}{Character string specifying the type of residual to compute.
Current options include:
\itemize{
\item \code{"function"} - (the default) for a list of functional residuals;
\item \code{"surrogate"} for a sample of surrogate residuals;
\item \code{"probscale"} for probability-scale residuals.
}}

\item{link.scale}{Logical indicating whether or not surrogate residuals
(\code{type = "surrogate"}) should be returned on the link scale (\code{TRUE}) vs. the
probability scale (\code{FALSE}). Default is \code{TRUE}.}

\item{...}{Additional optional arguments. Currently ignored.}
}
\value{
Either a list of functions (\code{type = "function"}) that also inherits
from class \code{"funres"} or a numeric vector of residuals (\code{type = "surrogate"}
or \code{type = "probscale"}).
}
\description{
Computes the function residual described in \url{TBD}.
}
\examples{
# Generate data from a logistic regression model with quadratic form
set.seed(1217)
n <- 1000
x <- rnorm(n)
z <- 1 - 2*x + 3*x^2 + rlogis(n)
y <- ifelse(z > 0, 1, 0)

# Fit models with/without quadratic term
fit.wrong <- glm(y ~ x, family = binomial)  # wrong
fit.right <- glm(y ~ x + I(x^2), family = binomial)  # right

# Generate functional residuals
fres.wrong <- fresiduals(fit.wrong)
plot(fres.wrong[[1]])  # plot first functional residual

# Function-function plot
par(mfrow = c(1, 2))
ffplot(fres.wrong, type = "l")
ffplot(fit.wrong, type = "l")

# Residual vs. predictor plot for each model based on surrogate method
par(mfrow = c(1, 2), las = 1)
lpars <- list(col = 2, lwd = 2)
col <- adjustcolor(1, alpha.f = 0.1)
palette("Okabe-Ito")
scatter.smooth(x, y = fresiduals(fit.wrong, type = "surrogate"),
               lpars = lpars, col = col, main = "Wrong model",
               xlab = "x", ylab = "Surrogate residual")
abline(h = 0, col = 3, lty = 2)
scatter.smooth(x, y = fresiduals(fit.right, type = "surrogate"),
               lpars = lpars, col = col, main = "Correct model",
               xlab = "x", ylab = "Surrogate residual")
abline(h = 0, col = 3, lty = 2)
}
\references{
TODO: Reference paper when published in JASA.
}
