% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frplot.R
\name{frplot}
\alias{frplot}
\alias{frplot.fresiduals}
\title{Functional residual-vs-predictor plots}
\usage{
frplot(
  object,
  x,
  scale = c("uniform", "normal"),
  resolution = 101,
  ref = TRUE,
  smooth = TRUE,
  ref_linetype = "dashed",
  ref_color = "white",
  smooth_linetype = "solid",
  smooth_color = "red",
  ...
)

\method{frplot}{fresiduals}(
  object,
  x,
  scale = c("uniform", "normal"),
  resolution = 101,
  ref = TRUE,
  smooth = TRUE,
  ref_linetype = "dashed",
  ref_color = "white",
  smooth_linetype = "solid",
  smooth_color = "red",
  ...
)
}
\arguments{
\item{object}{An object of class \link{fresiduals}.}

\item{x}{Vector of predictor values.}

\item{scale}{Character string specifying which scale to use for plotting.
Default is \code{"uniform"}.}

\item{resolution}{Integer specifying the resolution of the plot. Default
is 101.}

\item{ref}{Logical indicating whether or not to include a horizonal
reference line. Default is \code{TRUE}.}

\item{smooth}{Logical indicating whether or not to include a LOESS smoother
on the plot. Default is \code{TRUE}.}

\item{ref_linetype}{Character string or integer specifying the line type to
use for the reference line.}

\item{ref_color}{Character string or integer specifying the color to
use for the reference line.}

\item{smooth_linetype}{Character string or integer specifying the line type
to use for the smoother.}

\item{smooth_color}{Character string or integer specifying the color to
use for the smoother.}

\item{...}{Additional optional arguments. (Currently ignored.)}
}
\description{
Generates a functional residual-vs-predictor plot, visualizing the density of
the functional residuals as a heatmap rendered on either the uniform or
normal scale.
}
