% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funres-package.R
\docType{package}
\name{funres-package}
\alias{funres}
\alias{funres-package}
\title{funres: Functional Residuals}
\description{
An implementation of the functional residual framework described in XYZ. Functional residuals provide a novel diagnostic framework for generalized linear models (GLMs) and beyond, addressing limitations in traditional residuals for discrete outcome data. The framework uses functional residuals, which capture residual randomness that classical point statistics fail to represent, and provides new diagnostic tools like the functional-residual-vs-covariate and function-to-function plots. It broadens the diagnostic scope across various GLMs, including models for binary, ordinal, count data, and semiparametric models, offering a unified approach for both discrete and continuous data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bgreenwell.github.io/funres/}
}

}
\author{
\strong{Maintainer}: Brandon M. Greenwell \email{greenwell.brandon@gmail.com} (\href{https://orcid.org/0000-0002-8120-0084}{ORCID})

Authors:
\itemize{
  \item Zewei Lin \email{pqt19@txstate.edu}
  \item Brad Boehmke \email{bradleyboehmke@gmail.com} (\href{https://orcid.org/0000-0002-3611-8516}{ORCID})
  \item Dungang Liu \email{dungang.liu@uc.edu}
}

}
\keyword{internal}
