% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fresiduals.R
\name{fresiduals}
\alias{fresiduals}
\alias{fresiduals.vglm}
\title{Functional residuals}
\usage{
fresiduals(object, ...)

\method{fresiduals}{vglm}(object, ...)
}
\arguments{
\item{object}{An object of class \link[VGAM:vglm]{vglm}.}

\item{...}{Additional optional arguments. (Currently ignored.)}
}
\value{
An \code{n}-by-2 matrix...
}
\description{
Computes the function residual for XYZ as described in XYZ.
}
