% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funres-package.R
\docType{package}
\name{funres-package}
\alias{funres}
\alias{funres-package}
\title{funres: Functional Residuals}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

}
\keyword{internal}
